fix: 수정된 DB 스키마에 맞게 entity 코드 수정
